+++ core/src/test/java/org/testcontainers/junit/BaseDockerComposeTest.java
@@ -39,9 +39,6 @@ public abstract class BaseDockerComposeTest {
     public void simpleTest() {
         Jedis jedis = new Jedis(getEnvironment().getServiceHost("redis_1", REDIS_PORT), getEnvironment().getServicePort("redis_1", REDIS_PORT));
 
-        // TODO: remove following resolution of #160
-        Unreliables.retryUntilSuccess(10, TimeUnit.SECONDS, getLivenessCheck(jedis));
-
         jedis.incr("test");
         jedis.incr("test");
         jedis.incr("test");
@@ -53,9 +50,6 @@ public abstract class BaseDockerComposeTest {
     public void secondTest() {
         // used in manual checking for cleanup in between tests
         Jedis jedis = new Jedis(getEnvironment().getServiceHost("redis_1", REDIS_PORT), getEnvironment().getServicePort("redis_1", REDIS_PORT));
-
-        // TODO: remove following resolution of #160
-        Unreliables.retryUntilSuccess(10, TimeUnit.SECONDS, getLivenessCheck(jedis));
 
         jedis.incr("test");
         jedis.incr("test");